fix: cargo lint fix

+++ pkgs/halo2-binary-merkle-tree/src/binary_merkle_tree.rs
@@ -242,9 +242,7 @@ mod tests {
 
         let depth = 10;
         let num_leaves = 1 << depth;
-        let leaves = (0..num_leaves)
-            .map(|i| F::from(i as u64))
-            .collect::<Vec<F>>();
+        let leaves = (0..num_leaves).map(|i| F::from(i)).collect::<Vec<F>>();
 
         // Insert leaves
         for leaf in leaves.iter() {
